rename and move stuff, add script to generate connector json files
