+++ .gitignore
@@ -1,5 +1,7 @@
 .vscode*
 scratch*
+actual*.json
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
