+++ confluent/example_rabbitmq.json
