+++ example.env
@@ -1,0 +1,17 @@
+ORACLE_USERNAME=<my_admin_user>
+ORACLE_PASSWORD=<my_admin_password>
+ORACLE_ENDPOINT=db-mod-demo.<random_characters>.us-west-2.rds.amazonaws.com
+ORACLE_PORT=1521
+
+CLOUDAMQP_URL=amqps://<my_virtual_host>:<my_rabbitmq_password>@beaver.rmq.cloudamqp.com/<my_virtual_host>
+CLOUDAMQP_VIRTUAL_HOST=<random_characters>
+CLOUDAMQP_PASSWORD=<my_rabbitmq_password>
+CLOUDAMQP_ENDPOINT=beaver-01.rmq.cloudamqp.com
+
+CCLOUD_API_KEY=<random_characters>
+CCLOUD_API_SECRET=<random_characters>
+
+MONGO_USERNAME=<my_mongo_username>
+MONGO_PASSWORD=<my_mongo_password>
+MONGO_ENDPOINT=<my_database>.<random>.mongodb.net
+MONGO_DATABASE_NAME=demo-database-modernization
