+++ oracle/prepare_database.py
@@ -1,5 +1,5 @@
 import oracledb
-import oracle_config
+import config
 from pathlib import Path
 
 def execute_file(connection: oracledb.Connection, sql_file: str) -> None:
@@ -27,10 +27,10 @@ if __name__=="__main__":
     sql_populate_database = Path(__file__).absolute().with_name("populate_database.sql")
     sql_enable_cdc = Path(__file__).absolute().with_name("enable_cdc.sql")
     with oracledb.connect(
-        user=oracle_config.username,
-        password=oracle_config.password,
-        dsn=oracle_config.dsn,
-        port=oracle_config.port) as connection:
+        user=config.username,
+        password=config.password,
+        dsn=config.dsn,
+        port=config.port) as connection:
 
         print("Populating CUSTOMERS table")
         execute_file(connection, sql_populate_database)
